Tidied up ready for PR and added examples to demo.py
